+++ saleor/tests/fixtures.py
@@ -89,8 +89,7 @@ from ..shipping.models import (
     ShippingMethodType,
     ShippingZone,
 )
-from ..site import AuthenticationBackends
-from ..site.models import AuthorizationKey, SiteSettings
+from ..site.models import SiteSettings
 from ..warehouse.models import Allocation, Stock, Warehouse
 from ..webhook.event_types import WebhookEventType
 from ..webhook.models import Webhook
@@ -340,7 +339,6 @@ def other_shipping_method(shipping_zone, channel_USD):
 
 @pytest.fixture
 def checkout_without_shipping_required(checkout, product_without_shipping):
-    checkout = checkout
     variant = product_without_shipping.variants.get()
     add_variant_to_checkout(checkout, variant, 1)
     checkout.save()
@@ -723,7 +721,12 @@ def shipping_method_channel_PLN(shipping_zone, channel_PLN):
 @pytest.fixture
 def color_attribute(db):  # pylint: disable=W0613
     attribute = Attribute.objects.create(
-        slug="color", name="Color", type=AttributeType.PRODUCT_TYPE
+        slug="color",
+        name="Color",
+        type=AttributeType.PRODUCT_TYPE,
+        filterable_in_storefront=True,
+        filterable_in_dashboard=True,
+        available_in_grid=True,
     )
     AttributeValue.objects.create(attribute=attribute, name="Red", slug="red")
     AttributeValue.objects.create(attribute=attribute, name="Blue", slug="blue")
@@ -733,7 +736,12 @@ def color_attribute(db):  # pylint: disable=W0613
 @pytest.fixture
 def color_attribute_without_values(db):  # pylint: disable=W0613
     return Attribute.objects.create(
-        slug="color", name="Color", type=AttributeType.PRODUCT_TYPE
+        slug="color",
+        name="Color",
+        type=AttributeType.PRODUCT_TYPE,
+        filterable_in_storefront=True,
+        filterable_in_dashboard=True,
+        available_in_grid=True,
     )
 
 
@@ -748,7 +756,12 @@ def pink_attribute_value(color_attribute):  # pylint: disable=W0613
 @pytest.fixture
 def size_attribute(db):  # pylint: disable=W0613
     attribute = Attribute.objects.create(
-        slug="size", name="Size", type=AttributeType.PRODUCT_TYPE
+        slug="size",
+        name="Size",
+        type=AttributeType.PRODUCT_TYPE,
+        filterable_in_storefront=True,
+        filterable_in_dashboard=True,
+        available_in_grid=True,
     )
     AttributeValue.objects.create(attribute=attribute, name="Small", slug="small")
     AttributeValue.objects.create(attribute=attribute, name="Big", slug="big")
@@ -758,7 +771,12 @@ def size_attribute(db):  # pylint: disable=W0613
 @pytest.fixture
 def weight_attribute(db):
     attribute = Attribute.objects.create(
-        slug="material", name="Material", type=AttributeType.PRODUCT_TYPE
+        slug="material",
+        name="Material",
+        type=AttributeType.PRODUCT_TYPE,
+        filterable_in_storefront=True,
+        filterable_in_dashboard=True,
+        available_in_grid=True,
     )
     AttributeValue.objects.create(attribute=attribute, name="Cotton", slug="cotton")
     AttributeValue.objects.create(
@@ -768,9 +786,49 @@ def weight_attribute(db):
 
 
 @pytest.fixture
+def file_attribute(db):
+    attribute = Attribute.objects.create(
+        slug="image",
+        name="Image",
+        type=AttributeType.PRODUCT_TYPE,
+        input_type=AttributeInputType.FILE,
+    )
+    AttributeValue.objects.create(
+        attribute=attribute,
+        name="test_file.txt",
+        slug="test_filetxt",
+        file_url="http://mirumee.com/test_media/test_file.txt",
+        content_type="text/plain",
+    )
+    AttributeValue.objects.create(
+        attribute=attribute,
+        name="test_file.jpeg",
+        slug="test_filejpeg",
+        file_url="http://mirumee.com/test_media/test_file.jpeg",
+        content_type="image/jpeg",
+    )
+    return attribute
+
+
+@pytest.fixture
+def file_attribute_with_file_input_type_without_values(db):
+    return Attribute.objects.create(
+        slug="image",
+        name="Image",
+        type=AttributeType.PRODUCT_TYPE,
+        input_type=AttributeInputType.FILE,
+    )
+
+
+@pytest.fixture
 def size_page_attribute(db):
     attribute = Attribute.objects.create(
-        slug="page-size", name="Page size", type=AttributeType.PAGE_TYPE
+        slug="page-size",
+        name="Page size",
+        type=AttributeType.PAGE_TYPE,
+        filterable_in_storefront=True,
+        filterable_in_dashboard=True,
+        available_in_grid=True,
     )
     AttributeValue.objects.create(attribute=attribute, name="10", slug="10")
     AttributeValue.objects.create(attribute=attribute, name="15", slug="15")
@@ -780,7 +838,12 @@ def size_page_attribute(db):
 @pytest.fixture
 def tag_page_attribute(db):
     attribute = Attribute.objects.create(
-        slug="tag", name="tag", type=AttributeType.PAGE_TYPE
+        slug="tag",
+        name="tag",
+        type=AttributeType.PAGE_TYPE,
+        filterable_in_storefront=True,
+        filterable_in_dashboard=True,
+        available_in_grid=True,
     )
     AttributeValue.objects.create(attribute=attribute, name="About", slug="about")
     AttributeValue.objects.create(attribute=attribute, name="Help", slug="help")
@@ -797,6 +860,31 @@ def author_page_attribute(db):
     )
     AttributeValue.objects.create(
         attribute=attribute, name="Test author 2", slug="test-author-2"
+    )
+    return attribute
+
+
+@pytest.fixture
+def page_file_attribute(db):
+    attribute = Attribute.objects.create(
+        slug="image",
+        name="Image",
+        type=AttributeType.PAGE_TYPE,
+        input_type=AttributeInputType.FILE,
+    )
+    AttributeValue.objects.create(
+        attribute=attribute,
+        name="test_file.txt",
+        slug="test_filetxt",
+        file_url="http://mirumee.com/test_media/test_file.txt",
+        content_type="text/plain",
+    )
+    AttributeValue.objects.create(
+        attribute=attribute,
+        name="test_file.jpeg",
+        slug="test_filejpeg",
+        file_url="http://mirumee.com/test_media/test_file.jpeg",
+        content_type="image/jpeg",
     )
     return attribute
 
@@ -1013,6 +1101,14 @@ def product(product_type, category, warehouse, channel_USD):
 
 
 @pytest.fixture
+def product_with_collections(
+    product, published_collection, unpublished_collection, collection
+):
+    product.collections.add(*[published_collection, unpublished_collection, collection])
+    return product
+
+
+@pytest.fixture
 def product_available_in_many_channels(product, channel_PLN):
     ProductChannelListing.objects.create(
         product=product, channel=channel_PLN, is_published=True,
@@ -1139,6 +1235,49 @@ def product_with_variant_with_two_attributes(
     )
     associate_attribute_values_to_instance(
         variant, size_attribute, size_attribute.values.first()
+    )
+
+    return product
+
+
+@pytest.fixture
+def product_with_variant_with_file_attribute(
+    color_attribute, file_attribute, category, warehouse, channel_USD
+):
+    product_type = ProductType.objects.create(
+        name="Type with variant and file attribute",
+        slug="type-with-file-attribute",
+        has_variants=True,
+        is_shipping_required=True,
+    )
+    product_type.variant_attributes.add(file_attribute)
+
+    product = Product.objects.create(
+        name="Test product with variant and file attribute",
+        slug="test-product-with-variant-and-file-attribute",
+        product_type=product_type,
+        category=category,
+    )
+    ProductChannelListing.objects.create(
+        product=product,
+        channel=channel_USD,
+        is_published=True,
+        currency=channel_USD.currency_code,
+        visible_in_listings=True,
+        available_for_purchase=datetime.date(1999, 1, 1),
+    )
+
+    variant = ProductVariant.objects.create(product=product, sku="prodVarTest",)
+    ProductVariantChannelListing.objects.create(
+        variant=variant,
+        channel=channel_USD,
+        price_amount=Decimal(10),
+        cost_price_amount=Decimal(1),
+        currency=channel_USD.currency_code,
+    )
+
+    associate_attribute_values_to_instance(
+        variant, file_attribute, file_attribute.values.first()
     )
 
     return product
@@ -1793,18 +1932,24 @@ def voucher_customer(voucher, customer_user):
 
 @pytest.fixture
 def order_line(order, variant):
-    channel_slug = order.channel.slug
-    net = variant.get_price(channel_slug)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    product = variant.product
+    channel = order.channel
+    channel_listing = variant.channel_listings.get(channel=channel)
+    net = variant.get_price(product, [], channel, channel_listing)
+    currency = net.currency
+    gross = Money(amount=net.amount * Decimal(1.23), currency=currency)
+    quantity = 3
+    unit_price = TaxedMoney(net=net, gross=gross)
     return order.lines.create(
-        product_name=str(variant.product),
+        product_name=str(product),
         variant_name=str(variant),
         product_sku=variant.sku,
         is_shipping_required=variant.is_shipping_required(),
-        quantity=3,
+        quantity=quantity,
         variant=variant,
-        unit_price=TaxedMoney(net=net, gross=gross),
-        tax_rate=23,
+        unit_price=unit_price,
+        total_price=unit_price * quantity,
+        tax_rate=Decimal("0.23"),
     )
 
 
@@ -1823,18 +1968,23 @@ def order_line_with_allocation_in_many_stocks(
         channel=channel_USD,
     )
 
-    channel_slug = order.channel.slug
-    net = variant.get_price(channel_slug)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    product = variant.product
+    channel_listing = variant.channel_listings.get(channel=channel_USD)
+    net = variant.get_price(product, [], channel_USD, channel_listing)
+    currency = net.currency
+    gross = Money(amount=net.amount * Decimal(1.23), currency=currency)
+    quantity = 3
+    unit_price = TaxedMoney(net=net, gross=gross)
     order_line = order.lines.create(
-        product_name=str(variant.product),
+        product_name=str(product),
         variant_name=str(variant),
         product_sku=variant.sku,
         is_shipping_required=variant.is_shipping_required(),
-        quantity=3,
+        quantity=quantity,
         variant=variant,
-        unit_price=TaxedMoney(net=net, gross=gross),
-        tax_rate=23,
+        unit_price=unit_price,
+        total_price=unit_price * quantity,
+        tax_rate=Decimal("0.23"),
     )
 
     Allocation.objects.bulk_create(
@@ -1862,17 +2012,23 @@ def order_line_with_one_allocation(
         channel=channel_USD,
     )
 
-    net = variant.get_price(order.channel.slug)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    product = variant.product
+    channel_listing = variant.channel_listings.get(channel=channel_USD)
+    net = variant.get_price(product, [], channel_USD, channel_listing)
+    currency = net.currency
+    gross = Money(amount=net.amount * Decimal(1.23), currency=currency)
+    quantity = 2
+    unit_price = TaxedMoney(net=net, gross=gross)
     order_line = order.lines.create(
-        product_name=str(variant.product),
+        product_name=str(product),
         variant_name=str(variant),
         product_sku=variant.sku,
         is_shipping_required=variant.is_shipping_required(),
-        quantity=2,
+        quantity=quantity,
         variant=variant,
-        unit_price=TaxedMoney(net=net, gross=gross),
-        tax_rate=Decimal(23),
+        unit_price=unit_price,
+        total_price=unit_price * quantity,
+        tax_rate=Decimal("0.23"),
     )
 
     Allocation.objects.create(
@@ -1928,7 +2084,7 @@ def order_with_lines(
         available_for_purchase=datetime.date.today(),
     )
     variant = ProductVariant.objects.create(product=product, sku="SKU_AA")
-    ProductVariantChannelListing.objects.create(
+    channel_listing = ProductVariantChannelListing.objects.create(
         variant=variant,
         channel=channel_USD,
         price_amount=Decimal(10),
@@ -1938,17 +2094,21 @@ def order_with_lines(
     stock = Stock.objects.create(
         warehouse=warehouse, product_variant=variant, quantity=5
     )
-    net = variant.get_price(channel_USD.slug)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    net = variant.get_price(product, [], channel_USD, channel_listing)
+    currency = net.currency
+    gross = Money(amount=net.amount * Decimal(1.23), currency=currency)
+    quantity = 3
+    unit_price = TaxedMoney(net=net, gross=gross)
     line = order.lines.create(
         product_name=str(variant.product),
         variant_name=str(variant),
         product_sku=variant.sku,
         is_shipping_required=variant.is_shipping_required(),
-        quantity=3,
+        quantity=quantity,
         variant=variant,
-        unit_price=TaxedMoney(net=net, gross=gross),
-        tax_rate=23,
+        unit_price=unit_price,
+        total_price=unit_price * quantity,
+        tax_rate=Decimal("0.23"),
     )
     Allocation.objects.create(
         order_line=line, stock=stock, quantity_allocated=line.quantity
@@ -1968,7 +2128,7 @@ def order_with_lines(
         available_for_purchase=datetime.date.today(),
     )
     variant = ProductVariant.objects.create(product=product, sku="SKU_B")
-    ProductVariantChannelListing.objects.create(
+    channel_listing = ProductVariantChannelListing.objects.create(
         variant=variant,
         channel=channel_USD,
         price_amount=Decimal(20),
@@ -1979,17 +2139,21 @@ def order_with_lines(
         product_variant=variant, warehouse=warehouse, quantity=2
     )
 
-    net = variant.get_price(channel_USD.slug)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    net = variant.get_price(product, [], channel_USD, channel_listing)
+    currency = net.currency
+    gross = Money(amount=net.amount * Decimal(1.23), currency=currency)
+    unit_price = TaxedMoney(net=net, gross=gross)
+    quantity = 2
     line = order.lines.create(
         product_name=str(variant.product),
         variant_name=str(variant),
         product_sku=variant.sku,
         is_shipping_required=variant.is_shipping_required(),
-        quantity=2,
+        quantity=quantity,
         variant=variant,
-        unit_price=TaxedMoney(net=net, gross=gross),
-        tax_rate=23,
+        unit_price=unit_price,
+        total_price=unit_price * quantity,
+        tax_rate=Decimal("0.23"),
     )
     Allocation.objects.create(
         order_line=line, stock=stock, quantity_allocated=line.quantity
@@ -2044,7 +2208,7 @@ def order_with_lines_channel_PLN(
         available_for_purchase=datetime.date.today(),
     )
     variant = ProductVariant.objects.create(product=product, sku="SKU_A_PLN")
-    ProductVariantChannelListing.objects.create(
+    channel_listing = ProductVariantChannelListing.objects.create(
         variant=variant,
         channel=channel_PLN,
         price_amount=Decimal(10),
@@ -2054,17 +2218,21 @@ def order_with_lines_channel_PLN(
     stock = Stock.objects.create(
         warehouse=warehouse, product_variant=variant, quantity=5
     )
-    net = variant.get_price(channel_PLN.slug)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    net = variant.get_price(product, [], channel_PLN, channel_listing)
+    currency = net.currency
+    gross = Money(amount=net.amount * Decimal(1.23), currency=currency)
+    quantity = 3
+    unit_price = TaxedMoney(net=net, gross=gross)
     line = order.lines.create(
         product_name=str(variant.product),
         variant_name=str(variant),
         product_sku=variant.sku,
         is_shipping_required=variant.is_shipping_required(),
-        quantity=3,
+        quantity=quantity,
         variant=variant,
-        unit_price=TaxedMoney(net=net, gross=gross),
-        tax_rate=23,
+        unit_price=unit_price,
+        total_price=unit_price * quantity,
+        tax_rate=Decimal("0.23"),
     )
     Allocation.objects.create(
         order_line=line, stock=stock, quantity_allocated=line.quantity
@@ -2084,7 +2252,7 @@ def order_with_lines_channel_PLN(
         available_for_purchase=datetime.date.today(),
     )
     variant = ProductVariant.objects.create(product=product, sku="SKU_B_PLN")
-    ProductVariantChannelListing.objects.create(
+    channel_listing = ProductVariantChannelListing.objects.create(
         variant=variant,
         channel=channel_PLN,
         price_amount=Decimal(20),
@@ -2095,17 +2263,21 @@ def order_with_lines_channel_PLN(
         product_variant=variant, warehouse=warehouse, quantity=2
     )
 
-    net = variant.get_price(channel_PLN.slug)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    net = variant.get_price(product, [], channel_PLN, channel_listing, None)
+    currency = net.currency
+    gross = Money(amount=net.amount * Decimal(1.23), currency=currency)
+    quantity = 2
+    unit_price = TaxedMoney(net=net, gross=gross)
     line = order.lines.create(
         product_name=str(variant.product),
         variant_name=str(variant),
         product_sku=variant.sku,
         is_shipping_required=variant.is_shipping_required(),
-        quantity=2,
+        quantity=quantity,
         variant=variant,
-        unit_price=TaxedMoney(net=net, gross=gross),
-        tax_rate=23,
+        unit_price=unit_price,
+        total_price=unit_price * quantity,
+        tax_rate=Decimal("0.23"),
     )
     Allocation.objects.create(
         order_line=line, stock=stock, quantity_allocated=line.quantity
@@ -2134,17 +2306,24 @@ def order_with_line_without_inventory_tracking(
     order, variant_without_inventory_tracking
 ):
     variant = variant_without_inventory_tracking
-    net = variant.get_price(order.channel.slug)
-    gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    product = variant.product
+    channel = order.channel
+    channel_listing = variant.channel_listings.get(channel=channel)
+    net = variant.get_price(product, [], channel, channel_listing)
+    currency = net.currency
+    gross = Money(amount=net.amount * Decimal(1.23), currency=currency)
+    quantity = 3
+    unit_price = TaxedMoney(net=net, gross=gross)
     line = order.lines.create(
         product_name=str(variant.product),
         variant_name=str(variant),
         product_sku=variant.sku,
         is_shipping_required=variant.is_shipping_required(),
-        quantity=3,
+        quantity=quantity,
         variant=variant,
-        unit_price=TaxedMoney(net=net, gross=gross),
-        tax_rate=23,
+        unit_price=unit_price,
+        total_price=unit_price * quantity,
+        tax_rate=Decimal("0.23"),
     )
 
     recalculate_order(order)
@@ -2405,21 +2584,6 @@ def discount_info(category, collection, sale, channel_USD):
         product_ids=set(),
         category_ids={category.id},  # assumes this category does not have children
         collection_ids={collection.id},
-    )
-
-
-@pytest.fixture
-def authorization_backend_name():
-    return AuthenticationBackends.FACEBOOK
-
-
-@pytest.fixture
-def authorization_key(site_settings, authorization_backend_name):
-    return AuthorizationKey.objects.create(
-        site_settings=site_settings,
-        name=authorization_backend_name,
-        key="Key",
-        password="Password",
     )
 
 
@@ -3083,17 +3247,6 @@ def fake_payment_interface(mocker):
 
 
 @pytest.fixture
-def mock_get_manager(mocker, fake_payment_interface):
-    mgr = mocker.patch(
-        "saleor.payment.gateway.get_plugins_manager",
-        autospec=True,
-        return_value=fake_payment_interface,
-    )
-    yield fake_payment_interface
-    mgr.assert_called_once()
-
-
-@pytest.fixture
 def staff_notification_recipient(db, staff_user):
     return StaffNotificationRecipient.objects.create(active=True, user=staff_user)
 
@@ -3199,8 +3352,11 @@ def allocation(order_line, stock):
 @pytest.fixture
 def allocations(order_list, stock, channel_USD):
     variant = stock.product_variant
-    net = variant.get_price(channel_USD)
+    product = variant.product
+    channel_listing = variant.channel_listings.get(channel=channel_USD)
+    net = variant.get_price(product, [], channel_USD, channel_listing)
     gross = Money(amount=net.amount * Decimal(1.23), currency=net.currency)
+    price = TaxedMoney(net=net, gross=gross)
     lines = OrderLine.objects.bulk_create(
         [
             OrderLine(
@@ -3211,8 +3367,9 @@ def allocations(order_list, stock, channel_USD):
                 variant_name=str(variant),
                 product_sku=variant.sku,
                 is_shipping_required=variant.is_shipping_required(),
-                unit_price=TaxedMoney(net=net, gross=gross),
-                tax_rate=23,
+                unit_price=price,
+                total_price=price,
+                tax_rate=Decimal("0.23"),
             ),
             OrderLine(
                 order=order_list[1],
@@ -3222,8 +3379,9 @@ def allocations(order_list, stock, channel_USD):
                 variant_name=str(variant),
                 product_sku=variant.sku,
                 is_shipping_required=variant.is_shipping_required(),
-                unit_price=TaxedMoney(net=net, gross=gross),
-                tax_rate=23,
+                unit_price=price,
+                total_price=price,
+                tax_rate=Decimal("0.23"),
             ),
             OrderLine(
                 order=order_list[2],
@@ -3233,8 +3391,9 @@ def allocations(order_list, stock, channel_USD):
                 variant_name=str(variant),
                 product_sku=variant.sku,
                 is_shipping_required=variant.is_shipping_required(),
-                unit_price=TaxedMoney(net=net, gross=gross),
-                tax_rate=23,
+                unit_price=price,
+                total_price=price,
+                tax_rate=Decimal("0.23"),
             ),
         ]
     )